src/test/kotlin/days/Day1Test.kt
package days

import kotlin.test.Test
import kotlin.test.assertEquals

class Day1Test : AdventOfCodeTest {

    private val day = Day1()
    private val inputFirst =
        """
        1abc2
        pqr3stu8vwx
        a1b2c3d4e5f
        treb7uchet
        """.trimIndent().split('\n')

    private val inputSecond =
        """
        two1nine
        eightwothree
        abcone2threexyz
        xtwone3four
        4nineeightseven2
        zoneight234
        7pqrstsixteen
        """.trimIndent().split('\n')

    @Test
    override fun solvePart1Test() {
        assertEquals("142", day.solvePart1(inputFirst))
    }

    @Test
    override fun solvePart2Test() {
        assertEquals("281", day.solvePart2(inputSecond))
    }
}